refactor to ESP adapter class
